requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pyTelegramBotAPI==4.8.0
python-dateutil==2.8.2
python-telegram-bot-calendar==1.0.5
requests==2.28.1
six==1.16.0
urllib3==1.26.13
